+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/environment-check
 
+## 0.16.20
+
+### Patch Changes
+
+-   Updated dependencies [de8a4878]
+    -   @sap-ux/axios-extension@1.9.0
+    -   @sap-ux/btp-utils@0.13.0
+
 ## 0.16.19
 
 ### Patch Changes