+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/telemetry
 
+## 0.2.3
+
+### Patch Changes
+
+-   Updated dependencies [de8a4878]
+    -   @sap-ux/btp-utils@0.13.0
+
 ## 0.2.2
 
 ### Patch Changes